+++ src/stuart-ml/linalg/DenseMatrix.lua
@@ -91,8 +91,9 @@ function DenseMatrix:index(i, j)
   end
 end
 
-function DenseMatrix:map()
-  error('NIY')
+function DenseMatrix:map(f)
+  local moses = require 'moses'
+  return DenseMatrix.new(self.numRows, self.numCols, moses.map(self.values, f), self.isTransposed)
 end
 
 function DenseMatrix:numActives()
@@ -138,7 +139,24 @@ function DenseMatrix:transpose()
   return DenseMatrix.new(self.numCols, self.numRows, self.values, not self.isTransposed)
 end
 
-function DenseMatrix:update(i, j, v)
+function DenseMatrix:update(...)
+  local moses = require 'moses'
+  local nargs = #moses.pack(...)
+  if nargs == 1 then
+    return self:updatef(...)
+  else
+    return self:update3(...)
+  end
+end
+
+function DenseMatrix:updatef(f)
+  for i=1,#self.values do
+    self.values[i] = f(self.values[i])
+  end
+  return self
+end
+
+function DenseMatrix:update3(i, j, v)
   self.values[self:index(i, j)] = v
 end