+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - [#43](https://github.com/BixData/stuart-ml/issues/43) `[util:]` Support loading LIBSVM files from `MLUtils`
+- `DenseMatrix` and `SparseMatrix` map() and update() support
 
 ## [2.0.1] - 2019-01-16
 ### Added